+++ influxdb_client/client/flux_table.py
@@ -3,7 +3,12 @@ Flux employs a basic data model built from basic data types.
 
 The data model consists of tables, records, columns.
 """
+import codecs
+import csv
+from http.client import HTTPResponse
 from json import JSONEncoder
+from typing import List, Iterator
+from influxdb_client.rest import _UTF_8_encoding
 
 
 class FluxStructure:
@@ -137,3 +142,145 @@ class FluxRecord(FluxStructure):
     def __repr__(self):
         """Format for inspection."""
         return f"<{type(self).__name__}: field={self.values.get('_field')}, value={self.values.get('_value')}>"
+
+
+class TableList(List[FluxTable]):
+    """:class:`~influxdb_client.client.flux_table.FluxTable` list with additionally functional to better handle of query result."""  # noqa: E501
+
+    def to_values(self, columns: List['str'] = None) -> List[List[object]]:
+        """
+        Serialize query results to a flattened list of values.
+
+        :param columns: if not ``None`` then only specified columns are presented in results
+        :return: :class:`~list` of values
+
+        Output example:
+
+        .. code-block:: python
+
+            [
+                ['New York', datetime.datetime(2022, 6, 7, 11, 3, 22, 917593, tzinfo=tzutc()), 24.3],
+                ['Prague', datetime.datetime(2022, 6, 7, 11, 3, 22, 917593, tzinfo=tzutc()), 25.3],
+                ...
+            ]
+
+        Configure required columns:
+
+        .. code-block:: python
+
+            from influxdb_client import InfluxDBClient
+
+                with InfluxDBClient(url="http://localhost:8086", token="my-token", org="my-org") as client:
+
+                # Query: using Table structure
+                tables = client.query_api().query('from(bucket:"my-bucket") |> range(start: -10m)')
+
+                # Serialize to values
+                output = tables.to_values(columns=['location', '_time', '_value'])
+                print(output)
+        """
+
+        def filter_values(record):
+            if columns is not None:
+                return [record.values.get(k) for k in columns]
+            return record.values.values()
+
+        return self._to_values(filter_values)
+
+    def to_json(self, columns: List['str'] = None, **kwargs) -> str:
+        """
+        Serialize query results to a JSON formatted :class:`~str`.
+
+        :param columns: if not ``None`` then only specified columns are presented in results
+        :return: :class:`~str`
+
+        The query results is flattened to array:
+
+        .. code-block:: javascript
+
+            [
+                {
+                    "_measurement": "mem",
+                    "_start": "2021-06-23T06:50:11.897825+00:00",
+                    "_stop": "2021-06-25T06:50:11.897825+00:00",
+                    "_time": "2020-02-27T16:20:00.897825+00:00",
+                    "region": "north",
+                     "_field": "usage",
+                    "_value": 15
+                },
+                {
+                    "_measurement": "mem",
+                    "_start": "2021-06-23T06:50:11.897825+00:00",
+                    "_stop": "2021-06-25T06:50:11.897825+00:00",
+                    "_time": "2020-02-27T16:20:01.897825+00:00",
+                    "region": "west",
+                     "_field": "usage",
+                    "_value": 10
+                },
+                ...
+            ]
+
+        The JSON format could be configured via ``**kwargs`` arguments:
+
+        .. code-block:: python
+
+            from influxdb_client import InfluxDBClient
+
+            with InfluxDBClient(url="http://localhost:8086", token="my-token", org="my-org") as client:
+
+                # Query: using Table structure
+                tables = client.query_api().query('from(bucket:"my-bucket") |> range(start: -10m)')
+
+                # Serialize to JSON
+                output = tables.to_json(indent=5)
+                print(output)
+
+        For all available options see - `json.dump <https://docs.python.org/3/library/json.html#json.dump>`_.
+        """
+        if 'indent' not in kwargs:
+            kwargs['indent'] = 2
+
+        def filter_values(record):
+            if columns is not None:
+                return {k: v for (k, v) in record.values.items() if k in columns}
+            return record.values
+
+        import json
+        return json.dumps(self._to_values(filter_values), cls=FluxStructureEncoder, **kwargs)
+
+    def _to_values(self, mapping):
+        return [mapping(record) for table in self for record in table.records]
+
+
+class CSVIterator(Iterator[List[str]]):
+    """:class:`Iterator[List[str]]` with additionally functional to better handle of query result."""
+
+    def __init__(self, response: HTTPResponse) -> None:
+        """Initialize ``csv.reader``."""
+        self.delegate = csv.reader(codecs.iterdecode(response, _UTF_8_encoding))
+
+    def __iter__(self):
+        """Return an iterator object."""
+        return self.delegate.__iter__()
+
+    def __next__(self):
+        """Retrieve the next item from the iterator."""
+        return self.delegate.__next__()
+
+    def to_values(self) -> List[List[str]]:
+        """
+        Serialize query results to a flattened list of values.
+
+        :return: :class:`~list` of values
+
+        Output example:
+
+        .. code-block:: python
+
+            [
+                ['New York', '2022-06-14T08:00:51.749072045Z', '24.3'],
+                ['Prague', '2022-06-14T08:00:51.749072045Z', '25.3'],
+                ...
+            ]
+        """
+        return list(self.delegate)